1.组织角色查询 2.组织角色添加 3.组织角色删除 4.组织角色修改 5.组织查询 6.组织添加

+++ src/main/java/evdc/vianet/auth/service/TeamServiceImp.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import evdc.vianet.auth.entity.Team;
+import evdc.vianet.auth.entity.TeamRole;
 import evdc.vianet.auth.mapper.TeamMapper;
 
 @Service("teamService")
@@ -20,17 +21,22 @@ public class TeamServiceImp implements TeamService {
 	}
 
 	@Override
-	public int deleteTeamById(int id) {
-		return teamMapper.deleteTeamById(id);
+	public int deleteTeamById(long id) {
+		Team team = teamMapper.findTeamById(id);
+		if(team.getDelete()==1) {
+			return 2;
+		}else {
+			return teamMapper.deleteTeamById(id);
+		}
 	}
 
 	@Override
-	public List<Team> findAll() {
+	public List<Team> findAllTeams() {
 		return teamMapper.findAllTeams();
 	}
 
 	@Override
-	public Team findTeamById(int id) {
+	public Team findTeamById(long id) {
 		// TODO 
 		return null;
 	}
